+++ src/components/Loader/Loader.tsx
@@ -1,1 +1,5 @@
-export const Loader = () => <div>Loading...</div>;
+import LoaderIcon from '@/assets/Loader.svg?react';
+
+import './Loader.scss';
+
+export const Loader = () => <LoaderIcon fill="#ABF8FF" />;
